feat: more modular approach at http requests
