+++ src/App.tsx
@@ -1,10 +1,12 @@
 import Spinner from './Components/Spinner/Spinner';
+import Toast from './Components/Toast/Toast';
 import WheelStructure from './Components/WheelStructure/WheelStructure';
 
 function App() {
   return (
     <>
       <Spinner />
+      <Toast />
       <WheelStructure />
     </>
   );
